+++ integrations/yotpo_shopper_loyalty/yotpo_shopper_loyalty/aws/store_return.py
@@ -112,7 +112,7 @@ def _create_yotpo_refund_request(return_data, refund_id):
     LOGGER.info(f'Creating yotpo refund request')
     refund_request['id'] = refund_id
     refund_request['order_id'] = return_data['order']['externalId']
-    refund_request['total_amount_cents'] = return_data['totalRefundAmount'] * 100
+    refund_request['total_amount_cents'] = int(float(return_data['totalRefundAmount']) * 100)
     refund_request['items'] = _get_returned_items(return_data['items']['nodes'])
     refund_request['currency'] = return_data['order']['currency']
     return refund_request
